🚚 Consolidate and move `DeprecationWarnings` module to `/lib`
