+++ config/environments/test.rb
@@ -1,6 +1,6 @@
 require "active_support/core_ext/integer/time"
 require "fileutils"
-require_relative "../../config/initializers/deprecation_warnings"
+require_relative "../../lib/deprecation_warnings"
 
 # The test environment is used exclusively to run your application's
 # test suite. You never need to work with it otherwise. Remember that
